Refactored TestFBConfigForWinlog and TestFBConfigForWinevtlog to include them in an unique GO file
